Added some missing comments and e-tag validation
Added some missing comments and e-tag validation for all kind of assets, that causes the browser to send a "If-None-Match" with a file-hash along the "If-Modified-Since" header. In general this makes caching way more flexible as the site editor do not have to wait until the cache expires to supply new content.